move mock_motor in test_motors.py
